requirements-dev.txt
pillow
inky
pygame

